Skill And Item Update

+++ Assets/0_MonkeySwing/PJH/Scripts/UI_Scripts_jh/Popup_manager.cs
@@ -87,7 +87,7 @@ public class Popup_manager : MonoBehaviour
 
     public void AddBanana()
     {
-        GameManagerEx.Instance.player.Money += BananaCount.bananacount;
+        GameManagerEx.Instance.player.Money += (int)BananaCount.bananacount;
         BananaCount.bananacount = 0;
     }
